Dynamic namespace support.
This commit changes how the Telepresence traffic-manager and mutating
webhook deals with namespaces so that everything is controlled by one
single namespace selector.

A selector can be _dynamic_ or _static_. This in turn controls if
telepresence is "cluster-wide" or "namespaced". A _dynamic_ selector
requires cluster-wide permissions for the traffic-manager, and only a
_static_ selector can serve as base when installing Role/RoleBinding
pairs.

A selector is considered static if it meets the following conditions:
- The selector must not have `matchLabels`.
- The selector must have exactly one element in the `matchLabels` or the
  `matchExpression` list (if the element is in the `matchLabels` list,
   it is normalized into "key in [value]").
- The element must meet the following criteria:
  - The `key` of the element must be "kubernetes.io/metadata.name".
  - The `operator` of the element must be "In" (case-sensitive).
  - The `values` list of the element must contain at least one value.

Signed-off-by: Thomas Hallgren <thomas@tada.se>