Let genversion pick the latest version from the CHANGELOG.yml file.
The `TELEPRESENCE_VERSION` defaults to the version generated by the
`genversion` command. Prior to this commit, this command would determine
the version from a call to `git describe --tags --match='v*'`, yielding
a version with the latest known major.minor numbers. This means that
when starting a new minor version, the old minor version would still be
used, unless an arbitrary tag was set to change this behavior.

This commit changes the `genversion` command so that it instead picks
the latest version from the CHANGELOG.md. It will still use the
`git describe --tags --match='v*'` command to get the git generated
pre-release hash.

Signed-off-by: Thomas Hallgren <thomas@tada.se>