Drop use of the telepresence-agents configmap.
This commit removes the complex pattern used when engaging with a
workload to get a traffic-agent injected into its pods.

1. Generate an agent configuration.
2. Store the configuration in the `telepresence-agents` configmap.
3. Wait for the agent to arrive.

A watcher on the `telepresence-agents` configmap would now get an event
and discover the new or modified configuration. It took the following
steps:
1. Determine if a rollout is needed.
2. Perform a rollout to trigger the mutating webhook.

1. Check if the pod has a configuration entry in the configmap.
2. Patch pod if necessary.

1. Generate an agent configuration.
2. Evict pods with config annotation mismatch (triggers the webhook).

No action. The configmap no longer exists

1. Check if the pod has a configuration entry in the configmap.
2. Patch pod if necessary and include the config as an annotation.

Signed-off-by: Thomas Hallgren <thomas@tada.se>