Using the --proxy-via flag would sometimes cause connection timeouts.
Typically, a `telepresence connect --proxy-via <subnet>=<workflow>`
would fail with a "deadline exceeded" message when several workloads
were present in the namespace, the one targeted by the proxy-via didn't
yet have an agent installed, and other workloads had an agent. This was
due to a race condition in the logic for the agent-based port-forwards
in the root daemon. The conditions causing this race are now eliminated.

Signed-off-by: Thomas Hallgren <thomas@tada.se>