Ensure that tunnel write-delays cause connection back-pressure.
When reading from a connection and then writing the data on a tunnel,
it is essential that any delay when writing also causes the next read
from the connection to be delayed. Delaying the connection read causes
a back-pressure on the TCP stack, and will make the sender aware that
things need to slow down, which is exactly what we want.

Closes #3766

Signed-off-by: Thomas Hallgren <thomas@tada.se>