Let daemon.WaitForIP wait for connection to become ready.
The agent that satisfies the IP is not ready to handle connection
attempts at the time when it shows up in the clients map, and this
will sometimes cause the sftp client to fail with a "connect: network is
unreachable" error.

This commit ensures that the port-forward is established before the
`WaitForIP` call returns.

Signed-off-by: Thomas Hallgren <thomas@tada.se>