Make the DNS recursion check configurable and turn it off by default.
Very few systems experience a DNS recursion lookup problem. It can only
occur when the cluster runs locally and the cluster's DNS is configured
to somehow use DNS server that is started by Telepresence. The check is
therefore now configurable through the client setting
`dns.recursionCheck`, and it is `false` by default.

Signed-off-by: Thomas Hallgren <thomas@tada.se>