Rename charts/telepresence to charts/telepresence-oss.
The Helm chart name "telepresence-oss" was inconsistent with its
contained folder "telepresence". As a result, attempts to install the
chart using an argo ApplicationSet failed. The contained folder was
renamed to match the chart name.

This commit renames the folder and also removes old code related to
"telepresence-crds". Telepresence OSS has no CRDs.

Closes #3801

Signed-off-by: Thomas Hallgren <thomas@tada.se>